+++ src/main/java/eggy/task/Task.java
@@ -1,6 +1,7 @@
+package eggy.task;
 public class Task {
-    protected String description;
-    protected boolean isDone;
+    public String description;
+    public boolean isDone;
 
     public Task(String description) {
         this.description = description;
