+++ src/main/java/eggy/task/ToDo.java
@@ -1,3 +1,4 @@
+package eggy.task;
 public class ToDo extends Task {
     public ToDo(String input) {
         super("");
